moved citation-box and util to new dir
